+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+* Added `dismissSearchBar` in `FlaneurSearchViewController` so that the search bar can be dismissed without the need of another view controller being presented
+
 ## [0.9.0] - 2018-02-20
 
 * Added `segmentedCollectionView(_:willDisplayHeader:)` to `SegmentedCollectionViewDelegate`