+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.10.0] - 2018-03-12
+
+* Added `dismissSearchBar` in `FlaneurSearchViewController` so that the search bar can be dismissed without the need of another view controller being presented
+* Refactored/renamed `ProgressModalViewController` so that it is easier to implement
+* Added a cancel button to `ProgressModalViewController`
+
 ## [0.9.0] - 2018-02-20
 
 * Added `segmentedCollectionView(_:willDisplayHeader:)` to `SegmentedCollectionViewDelegate`
@@ -125,6 +131,7 @@ Older versions of this package were evolving at a really high pace.
 The package was not stable enough and maintaining a Changelog would have a real
 pain in the b***.
 
+[0.10.0]: https://github.com/FlaneurApp/FlaneurOpen/compare/0.9.0...0.10.0
 [0.9.0]: https://github.com/FlaneurApp/FlaneurOpen/compare/0.8.0...0.9.0
 [0.8.0]: https://github.com/FlaneurApp/FlaneurOpen/compare/0.7.0...0.8.0
 [0.7.0]: https://github.com/FlaneurApp/FlaneurOpen/compare/0.6.0...0.7.0